# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ForceFlow API — Multi-Tenant Salesforce Automation Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team/invite:
    post:
      operationId: inviteTeamMember
      tags: [team]
      summary: Invite a new team member (Frontier only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInviteInput"
      responses:
        "201":
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamInvite"
  /team/{memberId}:
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update a team member's role
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantMember"
    delete:
      operationId: removeTeamMember
      tags: [team]
      summary: Remove a team member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  # ─── SALESFORCE ORGS ────────────────────────────────────────────────────────
...
  /pipelines:
    get:
      operationId: listPipelineRuns
      tags: [pipelines]
      summary: List pipeline runs for the tenant
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Pipeline runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineRun"
  /pipelines/{runId}:
    get:
      operationId: getPipelineRun
      tags: [pipelines]
      summary: Get a specific pipeline run with stage details
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pipeline run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRun"
  /pipelines/{runId}/approve:
    post:
      operationId: approvePipelineStage
      tags: [pipelines]
      summary: Approve a pipeline gate (Frontier only)
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineApprovalInput"
      responses:
        "200":
          description: Updated pipeline run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRun"

  # ─── QUALITY REPORTS ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TenantMember:
      type: object
      required: [id, tenantId, userId, role, joinedAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        userId:
          type: string
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [frontier, developer, sales, support, sme]
        invitedBy:
          type: ["string", "null"]
        joinedAt:
          type: string
          format: date-time

    # ─── ONBOARDING ────────────────────────────────────────────────────────
...
    TeamInviteInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [frontier, developer, sales, support, sme]
    TeamInvite:
      type: object
      required: [id, tenantId, email, role, token, createdAt, expiresAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        email:
          type: string
        role:
          type: string
        token:
          type: string
        acceptedAt:
          type: ["string", "null"]
          format: date-time
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    TeamMemberUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [frontier, developer, sales, support, sme]

    # ─── SALESFORCE ORGS ───────────────────────────────────────────────────
...
    PipelineRun:
      type: object
      required: [id, tenantId, status, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        gitConnectionId:
          type: ["string", "null"]
        triggerType:
          type: ["string", "null"]
        branch:
          type: ["string", "null"]
        commitSha:
          type: ["string", "null"]
        commitMessage:
          type: ["string", "null"]
        stages:
          type: ["object", "null"]
        status:
          type: string
          enum: [pending, running, passed, failed, cancelled]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
    PipelineApprovalInput:
      type: object
      required: [stage, decision]
      properties:
        stage:
          type: string
        decision:
          type: string
          enum: [approved, rejected]
        notes:
          type: string

    # ─── QUALITY ───────────────────────────────────────────────────────────
...